Extract state store setup
